Ecs sprite added!

+++ src/engine/rendering/window.h
@@ -4,7 +4,7 @@
 #include <GL/glew.h>
 #include <GLFW/glfw3.h>
 #include "../common/util.h"
-#include "../ecs/components/sprite.h"
+#include "../ecs/components/sprite_component.h"
 #include "shader.h"
 #include <stdbool.h>
 
@@ -13,17 +13,17 @@ typedef struct
 	GLFWwindow* handle;
 	u32 width, height;
 	float aspectRatio;
-
+    
 	bool has_background_image;
 	SpriteComponent backgroundImage;
-
+    
 } Window; Window window;
 
 static Shader background_shader;
 static GLint winAspectRatioLocation;
 
 void window_center();
-void window_create(u32 width, u32 height, const char* title, bool isFullscreen);
+void window_create(u16 width, u16 height, const char* title, bool isFullscreen);
 void window_setFunctions(void (*procInputs)(void), void (*upd)(void));
 
 void window_mainloop();
